feat: Enhance handleButtonInteraction to support Itinerary and Quotation types

+++ lib/utils.ts
@@ -33,39 +33,45 @@ function camelToSnake<T extends object>(obj: T): ConvertObjectKeysToSnake<T> {
   ) as ConvertObjectKeysToSnake<T>;
 }
 
-async function handleButtonInteraction<TData extends LeadsTableData>(
+async function handleButtonInteraction<
+  TData extends LeadsTableData | ItineraryTableData,
+>(
   action: "Delete" | "Add New",
-  type: "Lead" | "Itinary",
+  type: "Lead" | "Itinerary" | "Qoutation",
   selectedRows: Row<TData>[],
   table?: Table<TData>,
 ) {
   // Determine actions based on type and action
   if (type === "Lead") {
-    if (selectedRows.length) {
-      if (action === "Delete") {
-        const selectedLeadIds = selectedRows
-          .map((row) => row.original.leadId)
-          .filter(Boolean);
-        await deleteLeadByIds(selectedLeadIds);
-        table?.resetRowSelection();
-      }
+    if (action === "Delete" && selectedRows.length) {
+      const selectedLeadIds = selectedRows
+        .map((row) => {
+          const data = row.original as LeadsTableData;
+          return data.leadId;
+        })
+        .filter(Boolean);
+      await deleteLeadByIds(selectedLeadIds);
+      table?.resetRowSelection();
     }
     if (action === "Add New") {
       return ROUTES.NEW_LEADS;
     }
   }
-  // if (type === "Itinary") {
-  //   if (action === "Delete") {
-  //     const selectedItinaryIds = selectedRows
-  //       .map((row) => row.original.itenaryId)
-  //       .filter(Boolean);
-  //     await deleteItinaryByIds(selectedItinaryIds);
-  //   }
-  //   if (action === "Add New") {
-  //     return ROUTES.NEW_ITINARIES;
-  //   }
-  // }
-  else {
+  if (type === "Itinerary") {
+    if (action === "Delete" && selectedRows.length) {
+      const selectedItineraryIds = selectedRows
+        .map((row) => {
+          const data = row.original as ItineraryTableData;
+          return data.itineraryId;
+        })
+        .filter(Boolean);
+      // await deleteItineraryByIds(selectedItineraryIds);
+      table?.resetRowSelection();
+    }
+    if (action === "Add New") {
+      return ROUTES.NEW_ITINERARY;
+    }
+  } else {
     console.log("Unsuported type", type);
   }
 }
